scraper/requirements.txt
pycoingecko==3.1.0
kafka-python==2.0.2
requests==2.31.0 